refactor: simplify login page by integrating LoginForm component
Replaced the existing login form implementation with a new LoginForm component wrapped in a Suspense fallback. This change enhances code readability and maintainability by separating concerns and reducing the complexity of the LoginPage component.
